User Controller 통합
Login, Logout, Register, Ajax Controller -> User Controller
