+++ SpringBoardTest/SpringBoardTest/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -25,7 +25,5 @@
 
 	
 	<context:component-scan base-package="com.test.springboard.*" />
-
-	
 	
 </beans:beans>
