+++ SpringBoardTest/SpringBoardTest/src/main/webapp/index.jsp
@@ -4,14 +4,16 @@
 <html>
 <head>
 <meta charset="Shift_JIS">
+
 <script type="text/javascript">
 	function goLoginPage()
 	{
-			location.href="login.do";
+	 		location.href="login.do";
 	}
 </script>
 
 <body onload="goLoginPage()">
+<h1>Hello</h1>
 </body>
 </head>
 </html>
